+++ src/Enumerations/Priority.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the JobRouter Client.
+ * https://github.com/brotkrueml/jobrouter-client
+ *
+ * Copyright (c) 2019-2022 Chris Müller
+ *
+ * For the full copyright and license information, please view the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\JobRouterClient\Enumerations;
+
+enum Priority: int
+{
+    case Low = 1;
+    case Normal = 2;
+    case High = 3;
+}